Move `unroll_for_loops` into `ark-ff-macros`
